cec: split out reconciler logic from manager
Currently, the reconcile and "manager" logic is combined in the
same struct and file. This doesn't allow for easy unit-testing
of the reconcile logic - at least not without having to re-test
the whole manager logic.

Therefore, this commit extracts the reconciler into its own struct
and file.

* reconciler: process k8s events (CEC, CCEC & LocalNode) by decide whether they
  are relevant for the local node
* manager: process actual CEC (apply its resources to the xDS server, configure L7LB, ...)
